Fixed illumination system to be more similar to foundry.
Fixed global shadow to simulate non illuminations areas.

Fixed line of sight shadows to be more efficient, and accurate using raycasting to create shadow regions which are then used in the polygonal shadow system.

Added portal system that allows for opening doors/windows, and added synchronization with fog of war/line of sight. This is loaded from the map file.

Fixed context menu to update size based on the number of items, and added context menu for portals/tokens. The tokens one is seperated between the GM and the player.